:bug: (Customer) Métodos do controller em classe errada.
Issues: #27
